uagent/pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "uagent"
version = "0.1.0"
description = "AI agent for ucore — programs ESP32 devices through Jupyter notebooks"
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.0",
    "pyyaml",
    "agent-client-protocol",
    "jupyter-ai-acp-client",
    "jupyter-ai-persona-manager",
]

[project.entry-points."jupyter_ai.personas"]
ucore-acp = "uagent.persona:UCorePersona"

[tool.setuptools]
packages = ["uagent"]

[tool.setuptools.package-data]
uagent = ["prompts/*.txt", "skills/*.yaml", "skills/*.txt", "static/*.svg"]
